:truck: move 'search' endpoint to 'search' group
