+++ backend/src/zac/core/api/views.py
@@ -48,7 +48,6 @@ from ..services import (
     get_zaak_eigenschappen,
     get_zaakobjecten,
     get_zaaktypen,
-    get_zaken_es,
 )
 from ..views.utils import filter_documenten_for_permissions, get_source_doc_versions
 from ..zaakobjecten import GROUPS, ZaakObjectGroup
@@ -67,7 +66,6 @@ from .serializers import (
     RelatedZaakSerializer,
     RolSerializer,
     SearchEigenschapSerializer,
-    SearchSerializer,
     ZaakDetailSerializer,
     ZaakDocumentSerializer,
     ZaakEigenschapSerializer,
@@ -520,34 +518,3 @@ class EigenschappenView(ListAPIView):
             )
 
         return eigenschappen
-
-
-class SearchViewSet(views.APIView):
-    authentication_classes = (authentication.SessionAuthentication,)
-    permission_classes = (permissions.IsAuthenticated,)
-    serializer_class = SearchSerializer
-
-    @extend_schema(responses=ZaakDetailSerializer(many=True))
-    def post(self, request, *args, **kwargs):
-        input_serializer = self.serializer_class(data=request.data)
-        input_serializer.is_valid(raise_exception=True)
-
-        zaken = self.perform_search(input_serializer.data)
-        zaak_serializer = ZaakDetailSerializer(zaken, many=True)
-
-        return Response(zaak_serializer.data)
-
-    def perform_search(self, data) -> List[Zaak]:
-        user_perms = UserPermissions(self.request.user)
-
-        if data.get("zaaktype"):
-            zaaktype_data = data.pop("zaaktype")
-            zaaktypen = get_zaaktypen(
-                user_perms,
-                catalogus=zaaktype_data["catalogus"],
-                omschrijving=zaaktype_data["omschrijving"],
-            )
-            data["zaaktypen"] = [zaaktype.url for zaaktype in zaaktypen]
-
-        zaken = get_zaken_es(user_perms, size=50, query_params=data)
-        return zaken
