+++ backend/src/zac/elasticsearch/drf_api/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
 
-from .views import GetZakenView
+from .views import GetZakenView, SearchViewSet
 
 urlpatterns = [
     path("zaken/autocomplete", GetZakenView.as_view(), name="zaken-search"),
+    path("zaken", SearchViewSet.as_view(), name="search"),
 ]
